+++ test/client.ts
@@ -340,6 +340,8 @@ describe('client', () => {
             networkError: null,
             graphQLErrors: null,
             forceFetch: false,
+            fetchMore: false,
+            paginationParameters: [],
             fragmentMap: {},
             returnPartialData: false,
             lastRequestId: 1,